move parser to own file
