+++ Dockerfile
@@ -14,7 +14,7 @@ RUN echo hey && git clone -b dev-comms https://github.com/datacamp/oil.git \
 RUN pip3 install protowhat
 
 ADD . shellwhat
-RUN cd shellwhat && pip3 install -r requirements.txt && pip3 install -e . && rm -rf shellwhat/tests/__pycache__
+RUN cd shellwhat && pip3 install -r requirements.txt && pip3 install -e .
 
 WORKDIR /shellwhat
